Safari-609.1.11.1/LayoutTests/http/tests/security/cross-frame-access-history-get-override-expected.txt

CONSOLE MESSAGE: line 57: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.

----- tests for getting a targetWindow's history object's functions which have custom overrides.  The desired behavior is for the targetWindow to return the builtin function, not the override -----

PASS: canGet('targetWindow.history.back') should be 'false' and is.
PASS: canGet('targetWindow.history.forward') should be 'false' and is.
PASS: canGet('targetWindow.history.go') should be 'false' and is.
